+++ s4e-web/src/app/views/map-view/state/product/product.model.ts
@@ -1,4 +1,9 @@
 import {Legend} from '../legend/legend.model';
+import {ActiveState, EntityState} from '@datorama/akita';
+import {yyyymmdd} from '../../../../utils/miscellaneous/date-utils';
+
+export const PRODUCT_MODE_QUERY_KEY = 'pmode'
+export const PRODUCT_MODE_FAVOURITE = 'favourite'
 
 export interface Product {
   id: number | undefined;
@@ -9,4 +14,43 @@ export interface Product {
   legend: Legend | null | undefined;
   layerName: string;
   favourite: boolean;
+}
+
+export interface ProductState extends EntityState<Product>, ActiveState<number> {
+    ui: {
+        loadedMonths: string[];
+        selectedDate: string;
+        selectedDay: number;
+        selectedYear: number;
+        selectedMonth: number;
+        availableDays: string[];
+    };
+    loaded: boolean;
+}
+
+export interface ProductUI {
+  isLoading: boolean;
+  isFavouriteLoading: boolean;
+}
+
+export interface ProductUIState extends EntityState<ProductUI> {
+}
+
+export function createProductState(state: Partial<ProductState> = {}): ProductState {
+    const now = new Date();
+    return {
+        active: null,
+        error: null,
+        loading: true,
+        ui: {
+            loadedMonths: [],
+            selectedDate: yyyymmdd(now),
+            selectedDay: now.getDate(),
+            selectedYear: now.getFullYear(),
+            selectedMonth: now.getMonth(),
+            availableDays: []
+        },
+        loaded: false,
+        ...state
+    };
 }