+++ s4e-web/src/app/views/map-view/state/product/product.query.ts
@@ -1,25 +1,64 @@
 import {Injectable} from '@angular/core';
-import {QueryEntity} from '@datorama/akita';
-import {ProductState, ProductStore} from './product.store';
-import {Product} from './product.model';
+import {EntityUIQuery, QueryEntity} from '@datorama/akita';
+import {ProductStore} from './product.store';
+import {Product, PRODUCT_MODE_FAVOURITE, PRODUCT_MODE_QUERY_KEY, ProductState, ProductUIState} from './product.model';
 import {combineLatest, Observable} from 'rxjs';
 import {IUILayer} from '../common.model';
 import {map} from 'rxjs/operators';
 import {SceneQuery} from '../scene/scene.query.service';
+import {RouterQuery} from '@datorama/akita-ng-router-store';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ProductQuery extends QueryEntity<ProductState, Product> {
-  constructor(protected store: ProductStore, private sceneQuery: SceneQuery) {
+  public readonly ui: EntityUIQuery<ProductUIState>;
+
+  constructor(protected store: ProductStore,
+              private routerQuery: RouterQuery,
+              private sceneQuery: SceneQuery) {
     super(store);
+    this.createUIQuery();
+  }
+
+  public selectFavourites(): Observable<Product[]> {
+    return this.selectAll().pipe(map(products => products.filter(p => p.favourite)));
+  }
+
+  public selectFavouritesCount(): Observable<number> {
+    return this.selectFavourites().pipe(map(products => products.length));
+  }
+
+  public selectAllFilteredAsUILayer(): Observable<IUILayer[]> {
+    return combineLatest(
+      this.routerQuery.selectQueryParams(PRODUCT_MODE_QUERY_KEY),
+      this.selectAllAsUILayer()
+    ).pipe(
+      map(([mode, layers]) => {
+        switch(mode) {
+          case PRODUCT_MODE_FAVOURITE: {
+            return layers.filter(layer => layer.favourite);
+          }
+          default: {
+            return layers;
+          }
+        }
+      })
+    );
   }
 
   public selectAllAsUILayer(): Observable<IUILayer[]> {
-    return combineLatest(this.selectAll(), this.selectActiveId())
+    return combineLatest(this.selectAll(), this.ui.selectAll(), this.selectActiveId())
       .pipe(
-        map(([products, activeId]) => products
-          .map(pt => ({cid: pt.id, caption: pt.displayName, active: pt.id === activeId, favourite: pt.favourite}))
+        map(([products, productsUi, activeId]) => products
+          .map((pt, i) => ({
+            cid: pt.id,
+            caption: pt.displayName,
+            active: pt.id === activeId,
+            favourite: pt.favourite,
+            isLoading: productsUi[i].isLoading,
+            isFavouriteLoading: productsUi[i].isFavouriteLoading
+          }))
         )
       );
   }
@@ -30,5 +69,10 @@ export class ProductQuery extends QueryEntity<ProductState, Product> {
 
   selectAvailableDates() {
     return this.select(state => state.ui.availableDays);
+  }
+
+  selectIsFavouriteMode(): Observable<boolean> {
+    return this.routerQuery.selectQueryParams(PRODUCT_MODE_QUERY_KEY)
+      .pipe(map(param => param === PRODUCT_MODE_FAVOURITE));
   }
 }